chore: apply stash — PandaScore config fields, pandascore source, package renames
- config: add DataSource, SeriesId fields; validate per data source
- sources: merge ParseLiquipediaMatches/Schedule into liquipedia.go, delete parser.go
- sources: add pandascore.go (PandaScore API client)
- store/match_results: use sources.ParseLiquipediaMatches, drop dead fetchMatchDataFromExternal
- app/api: drop dead FetchScheduledMatches (killed by PR #39)
- bot/utils: fix import api/format → tournament
- test files: fix old package paths (api/external, api/format → sources, tournament)
- sources/liquipedia_test: drop GetWikitext tests (moved to configure.go in PR #39)
