+++ config/config_test.go
@@ -22,9 +22,10 @@ func writeTemp(t *testing.T, contents string) string {
 	return path
 }
 
-func TestLoad_AllFields(t *testing.T) {
+func TestLoad_AllFields_Liquipedia(t *testing.T) {
 	path := writeTemp(t, `
 tournament_name = "MyEvent_2026"
+data_source = "liquipedia"
 page = "Foo/2026/Bar/Stage_1"
 round = "Stage_1"
 upcoming_only = true
@@ -34,15 +35,37 @@ test = true
 	cfg, err := Load(path)
 	assert.NoError(t, err)
 	assert.Equal(t, "MyEvent_2026", cfg.TournamentName)
+	assert.Equal(t, "liquipedia", cfg.DataSource)
 	assert.Equal(t, "Foo/2026/Bar/Stage_1", cfg.Page)
 	assert.Equal(t, "Stage_1", cfg.Round)
 	assert.True(t, cfg.UpcomingOnly)
 	assert.True(t, cfg.Test)
 }
 
-func TestLoad_DefaultBools(t *testing.T) {
+func TestLoad_AllFields_PandaScore(t *testing.T) {
+	path := writeTemp(t, `
+tournament_name = "MyEvent_2026"
+data_source = "pandascore"
+series_id = 10583
+round = "Stage_1"
+upcoming_only = true
+test = true
+`)
+
+	cfg, err := Load(path)
+	assert.NoError(t, err)
+	assert.Equal(t, "MyEvent_2026", cfg.TournamentName)
+	assert.Equal(t, "pandascore", cfg.DataSource)
+	assert.Equal(t, 10583, cfg.SeriesId)
+	assert.Equal(t, "Stage_1", cfg.Round)
+	assert.True(t, cfg.UpcomingOnly)
+	assert.True(t, cfg.Test)
+}
+
+func TestLoad_DefaultBoolsAndOptionalParams(t *testing.T) {
 	path := writeTemp(t, `
 tournament_name = "X"
+data_source = "liquipedia"
 page = "Y/Z"
 round = "Z"
 `)
@@ -55,6 +78,7 @@ round = "Z"
 
 func TestLoad_MissingTournamentName(t *testing.T) {
 	path := writeTemp(t, `
+data_source = "liquipedia"
 page = "Y/Z"
 round = "Z"
 `)
@@ -67,6 +91,7 @@ round = "Z"
 func TestLoad_MissingPage(t *testing.T) {
 	path := writeTemp(t, `
 tournament_name = "X"
+data_source = "liquipedia"
 round = "Z"
 `)
 
@@ -77,11 +102,36 @@ round = "Z"
 func TestLoad_MissingRound(t *testing.T) {
 	path := writeTemp(t, `
 tournament_name = "X"
+data_source = "liquipedia"
 page = "Y/Z"
 `)
 
 	_, err := Load(path)
 	assert.Error(t, err)
+}
+
+func TestLoad_MissingSeriesId(t *testing.T) {
+	path := writeTemp(t, `
+tournament_name = "X"
+data_source = "pandascore"
+round = "Z"
+`)
+
+	_, err := Load(path)
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "series_id")
+}
+
+func TestLoad_UnsupportedDataSource(t *testing.T) {
+	path := writeTemp(t, `
+tournament_name = "X"
+data_source = "unknown"
+round = "Z"
+`)
+
+	_, err := Load(path)
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "unsupported")
 }
 
 func TestLoad_FileNotFound(t *testing.T) {
